+++ disparityview/disparity_view.py
@@ -12,7 +12,7 @@ import open3d as o3d
 from tqdm import tqdm
 
 
-from depthview.zed_camerainfo import CameraParameter
+from disparityview.zed_camerainfo import CameraParameter
 
 
 def finitemax(depth: np.ndarray) -> float:
