+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools", "setuptools_scm"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "depthview"
+name = "disparirtyview"
 description = "depth npy file viewer"
 readme = "README.md"
 version = "0.0.1"
@@ -28,14 +28,14 @@ dev = [
     "black",
 ]
 [project.scripts]
-disparity_viewer = "depthview.disparity_view:disparity_viewer_main"
-zed_capture = "depthview.zed_capture:main"
+disparity_viewer = "disparirtyview.disparity_view:disparity_viewer_main"
+zed_capture = "disparirtyview.zed_capture:main"
 
 [project.urls]
 Repository = "https://github.com/katsunori-waragai/depth-view"
 
 [tool.setuptools]
-packages = ["depthview"]
+packages = ["disparirtyview"]
 
 [tool.black]
 line-length = 120
