+++ test/test_zed_camerainfo.py
@@ -2,7 +2,7 @@ import pyzed.sl as sl
 
 import sys
 
-from depthview.zed_camerainfo import get_width_height_fx_fy_cx_cy, get_baseline, CameraParameter
+from disparityview.zed_camerainfo import get_width_height_fx_fy_cx_cy, get_baseline, CameraParameter
 
 
 def test_get_baseline():
